Refactor some planet stuff out of main
